Fixed :bug: w/ the roll=False argument in loadexpt
The history=40 argument is used to determine how many temporal
frames to keep in the toeplitz window. The bug was that when the
roll keyword argument was set to False, the history was still used
to clip the response array. Therefore, when roll was set to False,
and history was set to a non-zero value, the spikes and stimulus
would be misaligned. The fix is that I removed the roll=False argument
and just don't roll the stimulus when history=0.