Refactor of the IO module
Splits Keras-specific monitoring of stuff from the more generic stuff in
the io.Monitor class. there is now a KerasMonitor subclass that
specifies the Keras specific stuff. The idea is that we can then later
create a GLMMonitor or some other kind of monitor to compare models
trained using different codebases.